move course config
